requirements_local.txt
# Python 3.9+ Required
# Local deployment minimal dependencies

# ========== GUI Framework ==========
PyQt5>=5.15.0

# ========== AI/LLM & API ==========
openai>=1.3.5
tiktoken>=0.5.0

# ========== Network & HTTP ==========
requests>=2.31.0
urllib3>=2.0.0

# ========== NLP & Text Processing ==========
jieba>=0.42.1

# ========== Data Processing ==========
numpy>=1.24.0

# ========== Configuration ==========
python-dotenv>=1.0.0
PyYAML>=6.0.0

# ========== PDF Processing ==========
PyPDF2>=3.0.0
